// frontend/src/openapi/compassClient/models/ParticipantDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ParticipantDto
 */
export interface ParticipantDto {
    /**
     * 
     * @type {number}
     * @memberof ParticipantDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ParticipantDto
     */
    name?: string;
}

/**
 * Check if a given object implements the ParticipantDto interface.
 */
export function instanceOfParticipantDto(value: object): boolean {
    return true;
}

export function ParticipantDtoFromJSON(json: any): ParticipantDto {
    return ParticipantDtoFromJSONTyped(json, false);
}

export function ParticipantDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ParticipantDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function ParticipantDtoToJSON(value?: ParticipantDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'name': value['name'],
    };
}